requirements.txt
psycopg2 >= 2.6.1
momoko >= 2.2.3
tornado >= 4.3
passlib >= 1.6.5
bcrypt >= 2.0.0
python-dateutil >= 2.5.3
